# Use threadpool for finding labels in chunk