- [ ] Figure out a good heuristic for when to use the threadpool

Great when we have lots of decent size chunks, particularly the NWM county groupby: 600ms -> 400ms.

```
| Change   | Before [627bf2b6] <main>   | After [7da43646] <threadpool>   |   Ratio | Benchmark (Parameter)                      |
|----------|----------------------------|---------------------------------|---------|--------------------------------------------|
| -        | 15.3±0.2ms                 | 7.93±0.03ms                     |    0.52 | cohorts.NWMMidwest.time_find_group_cohorts |
```